Continued major project restructuring and overhaul [19]
Adapted CMake files to new project structure.

CMake 3.11 is required because add_library() is called without any
sources, those are added later via the target_sources() command.

CMake 3.13 is required because it is treating relative paths as being
relative to the current source directory instead to the target to which
sources were being added.

CMake now fails early if any required components is not found.