Disabled ScaLAPACK usage in Elemental's svd() implementation
If Elemental's svd() is applied to a matrix A and the number of MPI processes is larger than
maximum(m(size(a)), n(size(a))), then Elemental throws a "std::runtime_error: pdgesvd exited with info=..".
This is because PDGESVD fails with INFO = MIN(M,N) + 1, if number of MPI processes is larger than MAX(M,N).