Increased no. of inner iterations in Elemental's SVD and clarify shortcomings of ScaLAPACK's PDGESVD
Depending on the numerical properties of the input matrix, ScaLAPACK's PDGESVD fails with INFO = MIN(M,N) + 1
and Elemental then throws a "std::runtime_error: pdgesvd exited with info=..". INFO > 0 means that DBDSQR did not
converge a.k.a. PDGESVD has detected heterogeneity by finding that eigenvalues were not identical across the
process grid. In this case, the accuracy of the results from PDGESVD cannot be guaranteed [1].

According to Demmel and Kahan (1990) algorithm convergence only depends on the matrix elements and values
computed by the DBDSQR algorithm, but not on the size or shape of the input matrix [3].

ScaLAPACK's DBDSQR has a internal variable MAXITR, that controls the maximum number of passes of the algorithm
through its inner loop. The algorithms stops (and so fails to converge) if the number of passes through the
inner loop exceeds MAXITR*N**2 [2].

MAXITR cannot be changed except though changing and recompiling DBDSQR's Fortran code so we use Elemental's SVD,
which allows changing this value through El::SVDCtrl<>::bidiagSVDCtrl.qrCtrl.maxIterPerVal variable.

Ref.:
[1] http://www.netlib.org/scalapack/explore-html/d2/de6/pdgesvd_8f_source.html
[2] http://www.netlib.org/lapack/lapack-3.1.1/html/dbdsqr.f.html
[3] http://www.netlib.org/lapack/lawnspdf/lawn03.pdf